prints the first 98 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	unsigned long long a = 1, b = 1, c;
+	unsigned long int a = 1, b = 1, c;
 	int i;
 
 	printf("%llu, %llu, ", a, b);
